Lesson-03_Seminar04_Task26. Enter natural number and return the number of it's digits.

+++ Seminar04_Task24/Program.cs
@@ -1,6 +1,6 @@
 ﻿// Lesson-03. Seminar04_Task24. Find sum of all natural numbers from 1 to N
 Console.Clear();
-int t; // ticks of the clock gentrator
+long t; // ticks of the clock gentrator
 
 string inputString()
 {
